Authentication added for create-update fo products and  orders

+++ pages/api/orders/[id].js
@@ -2,9 +2,10 @@ import Order from "../../../model/Order";
 import dbConnect from '../../../util/mongo'
 
 const handler = async( req, res ) => {
-    const { method, query: { id } } = req;
+    const { method, cookies, query: { id } } = req;
 
     dbConnect();
+    const token = cookies.token;
 
     if( 'GET' === method ) {
         try {
@@ -14,6 +15,9 @@ const handler = async( req, res ) => {
             return res.status( 500 ).json( error )
         }
     } else if( 'PUT' === method ) {
+        if( !token || token!== process.env.TOKEN ) {
+            return res.status(401).json( 'Not Authenticated' );
+        }
         try {
             const order = await Order.findByIdAndUpdate( id, req.body, { new: true } );
             return res.status( 201 ).json( order )
@@ -21,6 +25,9 @@ const handler = async( req, res ) => {
             return res.status( 500 ).json( error )
         }
     } else if( 'DELETE' === method ) {
+        if( !token || token!== process.env.TOKEN ) {
+            return res.status(401).json( 'Not Authenticated' );
+        }
         try {
             const order = await Order.findByIdAndDelete( id );
             return res.status( 200 ).json( order )
