+++ pages/api/products/[id].js
@@ -1,9 +1,11 @@
 import Product from "../../../model/Product";
-import axios from "axios";
 
 export default async function handler(req, res) {
 
-    const { query, method } = req;
+    const { query, method, cookies } = req;
+
+    const token = cookies.token;
+
     if( 'GET' === method ) {
         try {
             const product = await Product.findById( query.id );
@@ -19,6 +21,9 @@ export default async function handler(req, res) {
         //     res.status(500).json( error )
         // }
     } else if( 'DELETE' === method ) {
+        if( !token || token!== process.env.TOKEN ) {
+            return res.status(401).json( 'Not Authenticated' );
+        }
         try {
             await Product.findByIdAndDelete( query.id );
             res.status( 200 ).json( 'Product Deleted!' )
